Added ability to attach to existing containers and improved ROS1 interface (fixes #1) (#439)
* Added launch/connect functionality to ROS1 API

* Allowed shared workspaces to be optional to improve portability

* added attach method to App

+++ src/roswire/app/instance.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
 __all__ = ("AppInstance",)
 
-import contextlib
 import os
 import shutil
 import typing
 from types import TracebackType
-from typing import Iterator, Optional, Type
+from typing import Optional, Type
 
 import attr
 import dockerblade
@@ -39,20 +38,20 @@ class AppInstance:
         Provides access to a bash shell for this container.
     files: dockerblade.files.FileSystem
         Provides access to the filesystem for this container.
-    _host_workspace: str
-        The absolute path to the shared directory for this container's
+    _host_workspace: str, optional
+        The absolute path to the shared directory, if any, for this container's
         workspace on the host machine.
     _dockerblade: dockerblade.container.Container
         Provides access to the underlying Docker container.
     """
 
     _dockerblade: dockerblade.container.Container = attr.ib()
-    _host_workspace: str = attr.ib(repr=False)
     app: "App" = attr.ib()
     shell: dockerblade.shell.Shell = attr.ib(repr=False, init=False, eq=False)
     files: dockerblade.files.FileSystem = attr.ib(
         repr=False, init=False, eq=False
     )
+    _host_workspace: Optional[str] = attr.ib(repr=False, default=None)
 
     def __attrs_post_init__(self) -> None:
         dockerblade = self._dockerblade
@@ -127,39 +126,22 @@ class AppInstance:
             shell=self.shell, directory=directory, files=self.files
         )
 
-    @contextlib.contextmanager
-    def ros1(self, port: int = 11311) -> Iterator[ROS1]:
-        """
-        Launches a context-managed roscore inside the container.
-        Upon exiting the context, the ROS master (and its associated resources)
-        will be destroyed.
+    def ros1(self, port: int = 11311) -> ROS1:
+        """Provides access to ROS1 inside this application instance.
 
         Parameters
         ----------
         port: int, optional
-            The port inside the container on which roscore should run.
-
-        Yields
-        ------
-        ROS1
-            An interface to the launched ROS Master.
+            The port that should be used by the ROS Master.
         """
-        assert port > 1023
-        command = f"roscore -p {port}"
-        process = self.shell.popen(command)
-        try:
-            yield ROS1(
-                description=self.app.description,
-                shell=self.shell,
-                files=self.files,
-                ws_host=self._host_workspace,
-                ip_address=self.ip_address,
-                port=port,
-            )
-        finally:
-            process.terminate()
-            process.wait(2.0)
-            process.kill()
+        return ROS1(
+            description=self.app.description,
+            shell=self.shell,
+            files=self.files,
+            ws_host=self._host_workspace,
+            ip_address=self.ip_address,
+            port=port,
+        )
 
     @property
     def ros2(self) -> ROS2:
@@ -171,13 +153,15 @@ class AppInstance:
         self._dockerblade.remove()
 
         workspace = self._host_workspace
-        if os.path.exists(workspace):
+        if workspace and os.path.exists(workspace):
             logger.debug(f"destroying app instance directory: {workspace}")
             shutil.rmtree(workspace)
             logger.debug(f"destroyed app instance directory: {workspace}")
 
     def persist(
-        self, repo: Optional[str] = None, tag: Optional[str] = None
+        self,
+        repo: Optional[str] = None,
+        tag: Optional[str] = None
     ) -> DockerImage:
         """Persists this application instance to a Docker image.
